+++ packages/api/src/index.ts
@@ -26,6 +26,7 @@ export * from "./data/memberships";
 export * from "./data/notifications";
 export * from "./data/orderPhotos";
 export * from "./data/orders";
+export * from "./data/orderTracking";
 export * from "./data/perDiemTrips";
 export * from "./data/push";
 export * from "./data/rates";